refactor: Moved checkConditionals function to useForm hook
